all the changes and stuff

+++ Ares17_Group3/AbstractNPC.h
@@ -29,10 +29,10 @@ public:
 	//	virtual btRigidBody* addBoundingCapsule() = 0;
 	virtual void render(Model * modelData, glm::mat4 view, glm::mat4 proj, GLuint shader, Player *player) = 0;
 	virtual void modifyHealth(double newHp) = 0;
-	virtual void moveNpc(vertex* v, int speed) = 0;
+	virtual void moveNpc(vertex* v, float speed) = 0;
 	virtual queue<vertex*> findPath(AdjacencyList *adjList, int startId, int endId) = 0;
 	virtual bool update(Model* modelData, glm::mat4 view,
-		glm::mat4 proj, float dt, Grid* _g, Player* player, GLuint shader, int speed) = 0;
+		glm::mat4 proj, float dt, Grid* _g, Player* player, GLuint shader, float speed) = 0;
 	//	virtual bool findCollision() = 0;
 	virtual void setAttack(double atk) = 0;
 	virtual double getAttack() = 0;
